function/project3functionapp/requirements.txt
azure-functions==1.11.2
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
psycopg2-binary==2.9.3
requests==2.28.1
urllib3==1.26.12
